Fix(step1):와 val이랑 toString때문에 이럴수 있구나..
