+++ src/main/kotlin/lotto/domain/PurchasedLottoTickets.kt
@@ -3,7 +3,8 @@ package lotto.domain
 class PurchasedLottoTickets(
     val lottoTickets: List<LottoTicket>
 ) {
-    val size = lottoTickets.size
+    var size: Int = lottoTickets.size
+        private set
 
     override fun toString(): String {
         return lottoTickets.joinToString("\n")
