+++ src/test/kotlin/lotto/domain/LottoTicketDispenserTest.kt
@@ -12,17 +12,4 @@ internal class LottoTicketDispenserTest {
         assertThat(LottoTicketDispenser(LottoTicketGenerator()).getLottoTickets(inputMoeny).size)
             .isEqualTo(inputMoeny / LottoTicket.LOTTO_PRICE)
     }
-
-    @Test
-    fun `주어지는 전략으로 만들어진 로또 번호 확인`() {
-        // given
-        val lottoNumberGenerateStrategy =
-            LottoNumberGenerateStrategy { LottoTicket((7..12).map { LottoNumber.valueOf(it) }) }
-        val expectedTicket = LottoTicketGenerator().makeLottoTicket("7,8,9,10,11,12".split(","))
-        // when
-        val purchasedLottoTickets = LottoTicketDispenser(lottoNumberGenerateStrategy).getLottoTickets(1000L)
-        // than
-        assertThat(purchasedLottoTickets.lottoTickets[0].toString())
-            .isEqualTo(expectedTicket.toString())
-    }
 }
